+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Copy-paste error and actually updated to NMCP 1.6.1
 - Wrong units being shown in some benchmark graphs
 
+# Changed
+
+- Minor performance improvements on all platforms for `BitSink`
+- Minor performance improvements on all platforms for `BitSource`
+
 ## [1.10.4]
 
 ### Changed